phase2: fix uncommitted v1 test assertion (bf16 precision) + record the v1.1 negative iteration
The v1 exact-decode test fix (torch.equal on the bf16 downcast, since tl.dot
runs TF32 ~1.6e-3 not fp32-exact) was made and validated in the v1 session but
never git-added -- so HEAD's committed suite had one failing test. Committing
it now: 35/35.

v1.1 (autotune + tl.interleave byte-dedup + bf16-MMA prefill) regressed every
axis -- 34/35 tests, P-fid 0.6-0.8x -> 1.00x (bf16 rounding forfeits the
fidelity edge that is the kernel's whole point), decode ~2x slower, prefill
slower (25.9 vs 17.7ms). Reverted; v1 stands. Full negative + per-change root
cause in RESULTS-phase2-v1.1-negative.md so the bundle isn't retried blind.
Lesson: one change at a time vs the v1 baseline; keep fp32/TF32 inputs (P-fid
is dtype-bought and load-bearing).

+++ kernel/test_nf4_grouped.py
@@ -102,8 +102,12 @@ class TestDecodeExactness:
         eye = torch.eye(K, dtype=torch.bfloat16, device=dev)  # one-hot rows
         out = gemm_4bit_grouped(eye, B, A, [K], [0])
         ref = dequant_ref(B[0], A[0], N, K)  # [N, K]
-        # out[j] = W @ e_j -> out.T == W (one product per sum: exact)
-        assert torch.equal(out.t().to(torch.float32), ref)
+        # out[j] = W @ e_j -> out.T == W, one product per sum. tl.dot runs TF32
+        # (~1e-3 rel rounding, measured 1.6e-3), which sits below bf16's 2^-8
+        # resolution — so at the kernel's actual output precision the equality
+        # is EXACT, and any wrong code/nibble-order still lands on a different
+        # codebook value entirely.
+        assert torch.equal(out.t(), ref.to(torch.bfloat16))
 
 
 @cuda
